+++ lib/asset/algo_validation.go
@@ -11,7 +11,7 @@ func (a *NewAlgo) Validate() error {
 	return validation.ValidateStruct(a,
 		validation.Field(&a.Key, validation.Required, is.UUID),
 		validation.Field(&a.Name, validation.Required, validation.Length(1, 100)),
-		validation.Field(&a.Category, validation.In(AlgoCategory_ALGO_SIMPLE, AlgoCategory_ALGO_COMPOSITE, AlgoCategory_ALGO_AGGREGATE, AlgoCategory_ALGO_METRIC)),
+		validation.Field(&a.Category, validation.In(AlgoCategory_ALGO_SIMPLE, AlgoCategory_ALGO_COMPOSITE, AlgoCategory_ALGO_AGGREGATE, AlgoCategory_ALGO_METRIC, AlgoCategory_ALGO_PREDICT)),
 		validation.Field(&a.Description, validation.Required),
 		validation.Field(&a.Algorithm, validation.Required),
 		validation.Field(&a.Metadata, validation.By(validateMetadata)),