+++ server/standalone/migration/000025_add_predict_algo.up.sql
@@ -1,0 +1,9 @@
+SELECT execute($$
+
+    INSERT INTO algo_categories(category)
+    VALUES ('ALGO_PREDICT');
+
+$$) WHERE 'ALGO_PREDICT' NOT IN (
+    SELECT category
+    FROM algo_categories
+);